+++ packages/types/oauth.d.ts
@@ -1,0 +1,7 @@
+export type OAuthTokenPayload = {
+  userId?: number | null;
+  teamId?: number | null;
+  token_type: string;
+  scope: string[];
+  clientId: string;
+};
