+++ packages/ui/components/avatar/index.ts
@@ -2,4 +2,6 @@ export { Avatar } from "./Avatar";
 export { UserAvatar } from "./UserAvatar";
 export type { AvatarProps } from "./Avatar";
 export { AvatarGroup } from "./AvatarGroup";
+export { UserAvatarGroup } from "./UserAvatarGroup";
+export { UserAvatarGroupWithOrg } from "./UserAvatarGroupWithOrg";
 export type { AvatarGroupProps } from "./AvatarGroup";
