+++ packages/ui/components/badge/InfoBadge.tsx
@@ -1,7 +1,7 @@
-import { Tooltip } from "@calcom/ui";
-import { Info } from "@calcom/ui/components/icon";
+import { Info } from "../icon";
+import { Tooltip } from "../tooltip/Tooltip";
 
-export default function InfoBadge({ content }: { content: string }) {
+export function InfoBadge({ content }: { content: string }) {
   return (
     <>
       <Tooltip side="top" content={content}>
