+++ packages/ui/components/badge/index.ts
@@ -1,5 +1,5 @@
 export { Badge } from "./Badge";
 export { UpgradeTeamsBadge } from "./UpgradeTeamsBadge";
 export { UpgradeOrgsBadge } from "./UpgradeOrgsBadge";
-
+export { InfoBadge } from "./InfoBadge";
 export type { BadgeProps } from "./Badge";
